requirements.txt
debugpy 
numpy 
matplotlib 
scikit-image 
albumentations 
tqdm 
pandas 
torchmetrics 
tensorboard 
imagecodecs
transformers[torch]
pillow
pycocotools

timm
accelerate

opencv-python
onnxruntime
faster-coco-eval

datasets
wandb

ultralytics
ray[tune]
wandb